add filter for site names.

+++ server/public/controllers/controller.js
@@ -1,16 +1,16 @@
 var myApp = angular.module('myApp', []);
 myApp.controller('AppCtrl', ['$scope', '$http', function ($scope, $http) {
-    console.log('from the controller');
+    console.log("welcome to csbsb. this is Mo's first MEAN stack website");
     $scope.isEditing = false;
     $scope.isInfo = false;
     $scope.detailEditing = {}
 
     var refresh = function () {
         $http.get('/sitelist').then(function (response) {
-            console.log('i got the data', response);
+            //console.log('i got the data', response);
             $scope.sites = response.data;
             if (!$scope.isInfo)
-            $scope.site = undefined;
+                $scope.site = undefined;
         });
     };
 
@@ -18,41 +18,49 @@ myApp.controller('AppCtrl', ['$scope', '$http', function ($scope, $http) {
 
     $scope.addSite = function () {
         console.log($scope.site);
-        if ( !$scope.site || isNaN($scope.site.latitude) || isNaN($scope.site.longitude)) {
+        if (!$scope.site || isNaN($scope.site.latitude) || isNaN($scope.site.longitude)) {
             alert('填写站点名和经纬度');
             refresh();
         } else {
+            $scope.site.latitude = parseFloat($scope.site.latitude).toFixed(5);
+            $scope.site.longitude = parseFloat($scope.site.longitude).toFixed(5);
             $http.post('/sitelist', $scope.site).then(function (response) {
                 console.log(response.data);
                 refresh();
+                $scope.isEditing=!$scope.isEditing;
             });
         }
     };
 
     $scope.remove = function (id) {
-        console.log(id, 'to be removed');
-        $http.delete('/sitelist/' + id).then(function (response) {
+        if (confirm('确定删除台站吗？此操作不可恢复')){
+            console.log(id, 'to be removed');
+            $http.delete('/sitelist/' + id).then(function (response) {
             refresh();
-        })
+            })
+        }
+        
     };
 
-    $scope.edit = function (id) {
-        console.log(id, 'to be edit');
-        $http.get('/sitelist/' + id).then(function (response) {
-            $scope.site = response.data;
-            $scope.isEditing = true;
-        })
-    };
+    // $scope.edit = function (id) {
+    //     console.log(id, 'to be edit');
+    //     $http.get('/sitelist/' + id).then(function (response) {
+    //         $scope.site = response.data;
+    //         $scope.isEditing = true;
+    //     })
+    // };
 
     $scope.update = function () {
         console.log($scope.site._id);
-        for (item in $scope.detailEditing){
+        for (item in $scope.detailEditing) {
             $scope.detailEditing[item] = false;
         }
-        if (isNaN($scope.site.latitude) || isNaN($scope.site.longitude)) {
-            alert('Error');
+        if (isNaN($scope.site.latitude) || isNaN($scope.site.longitude) || !$scope.site.name) {
+            alert('确认名称和经纬度是否正确');
             refresh();
         } else {
+            $scope.site.latitude = parseFloat($scope.site.latitude).toFixed(5);
+            $scope.site.longitude = parseFloat($scope.site.longitude).toFixed(5);
             $http.put('/sitelist/' + $scope.site._id, $scope.site).then(function (response) {
                 $scope.isEditing = false;
                 refresh();
@@ -69,10 +77,10 @@ myApp.controller('AppCtrl', ['$scope', '$http', function ($scope, $http) {
     }
 
 
-    $scope.deselect = function () {
-        $scope.site = undefined;
-        $scope.isEditing = false;
-    }
+    // $scope.deselect = function () {
+    //     $scope.site = undefined;
+    //     $scope.isEditing = false;
+    // }
 
     $scope.returnToHome = function () {
         $scope.isInfo = false;
@@ -80,24 +88,29 @@ myApp.controller('AppCtrl', ['$scope', '$http', function ($scope, $http) {
     }
 
     $scope.addNewAttribute = function () {
-        if (!isNaN($scope.newAttrName) || ($scope.newAttrName in $scope.site) || !$scope.newAttrName){
+        if (!isNaN($scope.newAttrName) || ($scope.newAttrName in $scope.site) || !$scope.newAttrName) {
             alert('参数名无效');
             $scope.newAttrName = '';
         }
-        else{
-        $scope.site[$scope.newAttrName] = $scope.newAttrValue;
-        $scope.newAttrName = '';
-        $scope.newAttrValue = '';
-       }
+        else {
+            $scope.site[$scope.newAttrName] = $scope.newAttrValue;
+            $scope.newAttrName = '';
+            $scope.newAttrValue = '';
+        }
     }
 
     $scope.delAttribute = function (name) {
+        if (['name', 'latitude', 'longitude'].indexOf(name) != -1){
+            alert('不能移除名称或经纬度');
+        }else{
             delete $scope.site[name];
+        }
+        
     }
 
 }]);
 
-myApp.controller('MessageCtrl',['$scope', '$http', '$sce',function ($scope, $http, $sce) {
+myApp.controller('MessageCtrl', ['$scope', '$http', '$sce', function ($scope, $http, $sce) {
 
     var refresh = function () {
         $http.get('/message').then(function (response) {
@@ -108,21 +121,21 @@ myApp.controller('MessageCtrl',['$scope', '$http', '$sce',function ($scope, $htt
     refresh();
 
     $scope.addMsg = function () {
-        if ($scope.name!="" && $scope.msg!="" &&$scope.name &&$scope.msg){
-            $http.post('/message',{name:$scope.name, msg:$scope.msg}).then(function (response) {
+        if ($scope.name != "" && $scope.msg != "" && $scope.name && $scope.msg) {
+            $http.post('/message', {name: $scope.name, msg: $scope.msg}).then(function (response) {
                 $scope.name = '';
                 $scope.msg = '';
             });
             refresh();
-        }else{
+        } else {
             alert('填写名字')
         }
-       
+
     };
 
-    $scope.replaceText = function (str){
-    var reg=new RegExp("\n","g");    
-    str = str.replace(reg,"<br>");
-    return $sce.trustAsHtml(str);
+    $scope.replaceText = function (str) {
+        var reg = new RegExp("\n", "g");
+        str = str.replace(reg, "<br>");
+        return $sce.trustAsHtml(str);
     }
 }]);
